pyproject.toml
[project]
name = "gget-mcp"
version = "0.1.0"
description = "MCP server for gget bioinformatics library"
readme = "README.md"
authors = [
    { name = "antonkulaga", email = "antonkulaga@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "eliot>=1.17.5",
    "pycomfort>=0.0.18",
    "fastmcp>=2.7.0",
    "fastapi>=0.115.12",
    "gget>=0.29.1",
    "pandas>=2.0.0",
    "biopython>=1.84",
]

[project.scripts]
gget-mcp = "gget_mcp.server:cli_app_stdio"
server = "gget_mcp.server:cli_app"
stdio = "gget_mcp.server:cli_app_stdio"
sse = "gget_mcp.server:cli_app_sse"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gget_mcp"]

[tool.hatch.build]
exclude = [
    "*.log",
    "logs/",
    "images/",
    ".venv/",
    ".pytest_cache/",
    "test/",
    "*.tmp",
    "temp*",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "ruff>=0.8.0",
    "jupyter>=1.0.0",
    "just-agents-core>=0.7.4",
    "python-dotenv>=1.0.0",
] 
